code/Retrieval/services/answer_generation/v1.0.0/config.py
#!/usr/bin/env python3
"""
Configuration for Answer Generation Service v1.0.0
LLM-based answer generation with context and citations
"""

import os
import sys
from pathlib import Path
from dotenv import load_dotenv

# Load common .env from PipeLineServices root
env_path = Path(__file__).resolve().parents[4] / ".env"
load_dotenv(env_path)

# Add shared module to path
sys.path.insert(0, str(Path(__file__).resolve().parents[4]))
from shared import get_llm_for_task, LLMModels

# Service metadata
API_VERSION = "1.0.0"
SERVICE_NAME = "Answer Generation Service"
SERVICE_DESCRIPTION = "LLM-based answer generation with retrieved context and citations"

# Server config
DEFAULT_HOST = os.getenv("HOST", "0.0.0.0")
DEFAULT_PORT = int(os.getenv("ANSWER_SERVICE_PORT", "8074"))

# Dependent service URLs
# Note: Use base URL without endpoint path (answer_api.py adds /health for health checks and /v1/chat/completions for API calls)
LLM_GATEWAY_URL = os.getenv("LLM_GATEWAY_URL_DEVELOPMENT", "http://localhost:8065").replace("/v1/chat/completions", "")

# LLM parameters for answer generation
# Use shared registry for model selection (no more hardcoded models!)
# Default: Complex answer generation model from registry
DEFAULT_LLM_MODEL = os.getenv("DEFAULT_LLM_MODEL", get_llm_for_task("answer_generation", complexity="complex"))
DEFAULT_MAX_TOKENS = int(os.getenv("DEFAULT_MAX_TOKENS", "1024"))  # Reduced for speed
DEFAULT_TEMPERATURE = float(os.getenv("DEFAULT_TEMPERATURE", "0.2"))  # Lower for factual RAG answers

# Answer generation settings
MAX_CONTEXT_CHUNKS = int(os.getenv("MAX_CONTEXT_CHUNKS", "10"))  # Max chunks to include in context
ENABLE_CITATIONS = os.getenv("ENABLE_CITATIONS", "true").lower() == "true"
ENABLE_STREAMING = os.getenv("ENABLE_STREAMING", "true").lower() == "true"

# System prompt for answer generation (CITE ONLY RELEVANT SOURCES)
ANSWER_GENERATION_SYSTEM_PROMPT_RELEVANT_ONLY = """You are a knowledgeable assistant that provides accurate, detailed answers based on the given context.

Requirements:
- Base your answer ONLY on the provided context
- Be accurate and factual
- If the context doesn't contain enough information, say so clearly
- Include relevant details from the context
- Cite ONLY sources that contain information relevant to answering the question using [Source X] notation
- IMPORTANT: DO NOT mention, reference, or explain sources that are irrelevant or do not contribute to the answer
- IMPORTANT: If a source is not relevant, completely ignore it - do not say "other sources don't contain", "no additional information in other sources", or anything similar
- Only discuss and cite sources that you actually use in your answer
- Be concise but comprehensive
- Use clear, professional language

For comparison or analysis queries:
- CRITICAL: Look beyond exact terminology matches to identify semantic and conceptual similarities
- Analyze the PURPOSE and FUNCTION of features/technologies/attributes, not just their literal names
- Example: "cushioning" and "grip" serve analogous purposes across different domains (comfort, control, performance)
- Identify similarities, differences, patterns, or relationships across sources
- Synthesize insights beyond just listing isolated facts
- Look for thematic connections, shared concepts, functional equivalents, or contrasting approaches
- When identifying commonalities, be specific about WHAT is shared functionally, purposefully, or thematically
- If no exact terminology matches exist, explain what similar goals or functions are achieved through different means
- Provide analytical depth while staying grounded in the context
- Structure comparisons clearly: list each source's attributes, then analyze similarities/differences at both literal and conceptual levels

For cross-referencing queries (e.g., "which X appear in both Y and Z"):
- First, identify all items from category Y
- Second, identify all items from category Z
- Third, explicitly compare the two lists to find overlaps
- If NO overlap exists, state this clearly (e.g., "There is no X appearing in both Y and Z")
- Avoid assuming overlap when different items appear in different sources"""

# System prompt for answer generation (CITE ALL SOURCES)
ANSWER_GENERATION_SYSTEM_PROMPT_ALL_SOURCES = """You are a knowledgeable assistant that provides accurate, detailed answers based on the given context.

Requirements:
- Base your answer ONLY on the provided context
- Be accurate and factual
- If the context doesn't contain enough information, say so clearly
- Include relevant details from the context
- Cite sources using [Source X] notation when appropriate
- Explain which sources were relevant and which were not
- Be concise but comprehensive
- Use clear, professional language

For comparison or analysis queries:
- CRITICAL: Look beyond exact terminology matches to identify semantic and conceptual similarities
- Analyze the PURPOSE and FUNCTION of features/technologies/attributes, not just their literal names
- Example: "cushioning" and "grip" serve analogous purposes across different domains (comfort, control, performance)
- Identify similarities, differences, patterns, or relationships across sources
- Synthesize insights beyond just listing isolated facts
- Look for thematic connections, shared concepts, functional equivalents, or contrasting approaches
- When identifying commonalities, be specific about WHAT is shared functionally, purposefully, or thematically
- If no exact terminology matches exist, explain what similar goals or functions are achieved through different means
- Provide analytical depth while staying grounded in the context
- Structure comparisons clearly: list each source's attributes, then analyze similarities/differences at both literal and conceptual levels

For cross-referencing queries (e.g., "which X appear in both Y and Z"):
- First, identify all items from category Y
- Second, identify all items from category Z
- Third, explicitly compare the two lists to find overlaps
- If NO overlap exists, state this clearly (e.g., "There is no X appearing in both Y and Z")
- Avoid assuming overlap when different items appear in different sources"""

# Performance settings
REQUEST_TIMEOUT = int(os.getenv("REQUEST_TIMEOUT", "60"))  # seconds (longer for LLM)
ENABLE_CACHE = os.getenv("ENABLE_CACHE", "true").lower() == "true"
CACHE_TTL = int(os.getenv("CACHE_TTL", "300"))  # 5 minutes for fresh answers with evolving knowledge base

# Cache settings (Redis)
REDIS_HOST = os.getenv("REDIS_HOST", "localhost")
REDIS_PORT = int(os.getenv("REDIS_PORT", "6379"))
REDIS_DB = int(os.getenv("REDIS_DB", "3"))  # DB 3 for Answer Generation cache
