code/Retrieval/services/answer_generation/v1.0.0/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
pydantic==2.5.0
python-dotenv==1.0.0
redis==5.0.1
